refactor(local-mail): give Gmail message projection one owner
`headerValue` lived in schema.ts (wire schemas) and `bodyText` plus the
base64/MIME/HTML decoding lived in db.ts (the SQLite closure), but both do
one job: turn a `GmailMessage` wire object into the flat scalar fields the
mirror stores and serves. That invariant was split across two files by old
file boundaries, not by owner.

Move both into `message-fields.ts`, a pure `GmailMessage -> fields` module:
- schema.ts is now only TypeBox wire shapes, one sentence to explain.
- db.ts sheds ~58 lines of email-format decoding, keeping the SQLite handle
  lifecycle, ingest, and read models the prepared statements bind it to.
- The `GmailMessagePart` traversal type and the one cast off the loose
  `payload.parts` (`Type.Any()`) wire boundary now sit next to the code that
  reads them, so the cast reads as an intentional local boundary.

No behavior change: bodyText/header extraction are unchanged and still
covered by db.test.ts and read-models.test.ts through the read surface.
